+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add clear cache button to settings
 - Restrict local permissions for BitBoxApp config files and directories
 - Bitcoin: display zero amounts without decimal places
+- Add option for percentage on portfolio to reflect performance
 
 ## v4.51.3
 - Bundle BitBox02 and BitBox02 Nova firmware version v9.26.4